package/requirements.txt
audioread
cryptography
datasets
decorator
dictdiffer
dvc
dvc[s3]
GitPython
huggingface-hub
joblib
librosa
numpy
pandas
psycopg2-binary
pyarrow
PyYAML
requests
responses
rich
scikit-learn
scipy
scmrepo
soundfile
typer
typing_extensions
urllib3
